Moved filter by level to exclusive class
